msm: Kconfig: Add support for MSM_FIQ for 8625
On 8x25 modem is the master processor and in case apps
is holding lock with irqs disabled then sending an SMSM reset
interrupt does not help and the state of Apps is not known
at all.Hence we are enabling a FIQ which will save the cpu context
of the apps and help debugging before going into SMSM reset.
In the current patch we are adding support for enabling the fiq
and defining the fiq offset number.

Signed-off-by: Taniya Das <tdas@codeaurora.org>
(cherry picked from commit 87628c186a29a4454fb3097631085e15543f6bbb)

Signed-off-by: Gopal Goberu <ggober@codeaurora.org>

Conflicts:

	arch/arm/mach-msm/Kconfig

Change-Id: Icc9c059da56ce8e7910ea94ac35f7e651535bda5