CAF:QRD_AP:CAMERA:None:397699:msm: camera: Fix lens lock issue in actuator
	Component Name: CAMERA
	Feature Name: None
	Fixed CRs: 397699
	Targets: msm8X25_sku5
	Required VUs:
	Dependency Notes:
	Limitations:
	Suggested Testing: NONE
	Peer Review:
	Reference Build or VU:
	Comments: feizhang
	(*commit tool - 1.2.1)

The actuator is not able to move beyond one region boundary.
This is because when there is region boundary between current
lens position and destination lens position, the lens is first moved
to region boundary and then moved to the destination target position.
Once region boundary(target lens position in this case) is reached
we should not assume it is the end of total lens movement. So
condition check (curr_lens_pos == target_lens_pos) is wrong. Hence
removing the inaccurate condition check.

Change-Id: I13a313bf77d6aa236b74a010d92221d5b5ce360c